requirements.txt
playwright==1.57.0
pytest-playwright==0.7.2
python-dotenv==1.2.1
